requirements.txt

python-telegram-bot==20.7
yt-dlp>=2025.5.26